In progress work on firmware downloading.
Have added a new python utility program - audio_downloader.py which
turns a BINARY file into a wave file. If the -f option is given to it
then it outputs the firmware header on top of it.

Fixed up audio_test to handle the different stop definition and to ignore
the pre/post ambles. Also added the new DOWNLOAD_TYPE so downloads to Edison
will have that byte before the version.

TODO -- update tokens1.html with these formats for the audio downloads.