Pauses to all for flash erasure when programming firmware.
Also changed the audio_downloader.py program to audio_creator.py as
it couldn't actually use wx to play sounds. So now it just creates
the wav file which can be played with another utility.
Also made the output of audio_test.py better when it's converting
a wav file back to the binary that it came from.