Save/load program using Json.
All edison programs will now be saved as Json, and have the .edw extension. An old .mbw can be
still read in, but if saved it will be saved as a Json .edw file (so this can be used to
convert between the two formats).

Note that the load dialog now only shows .edw files by default, to find the .mbw files, show all
files first.

Redmine: #61