Most of not allowing a new event with the same check as the old one
Not all done, but almost. This code checks what's been done before presenting
the module and choice. Also don't allow for the event icon to be picked up
when there are no choices left.

Need a different colour icon for event, so that we can see if it's selected
for editing and if it's disabled because we have no choices left!