Fix for if-bric with Match# == 0
The code for matching an IR input, has to check two indications --
was there a match character received, and is the match character the one we are looking for.
The code treated a match for #0 the same as ANY match, so that's been fixed.

Also, now the status indication of a match is only cleared if the rest of the if is true
too. This will allow for multiple tests to be be done in sequence.