Added the ability to load ramp values from a file.
If a file "ramp_override.txt" is in the user's home directory, then
will read it looking for integer values between 0 and 100. They have
to be monotonically increasing. If any value is incorrect then the
default ramp is re-loaded. Not that this test is done before every
'program' attempt so the file can be changed in a single run of
edware.py.

The values are the percentage of the change (from the old value to the
new value) to apply at each sample. Therefore a changed can be 'ramped'
up (or down -- if the change is negative then the percentages still work).